Properly deal with stealth undefined command
If a valid but incomplete prefix sequence is followed by
an unbound key, then Emacs calls the `undefined' command
but does not set `this-command', `this-original-command'
or `real-this-command' accordingly.  Instead they are nil.

Deal with this by adding a corresponding entry to
`transient-predicate-map', replacing the old kludge
we used when we didn't know what was going on yet.