transient--post-command: Don't pop and push equal redisplay maps
`set-transient-map' arranges for the exit function not to be called
if another transient map was added later.  While we don't use that
function ourselves, this feature still backfires when we replace the
redisplay map, causing e.g. `mouse-drag-track's transient map to be
left in place when it should be removed.